Migrate from SDL2 to SDL3 using CMake FetchContent
- Replace bundled SDL2 libs with FetchContent for SDL3 3.4.4, SDL3_image 3.4.2, SDL3_ttf 3.2.2
- Update all source files from SDL2 API to SDL3 API:
  - SDL_Init, window/renderer creation, event handling, timing
  - SDL_RenderCopy -> SDL_RenderTexture, SDL_Rect -> SDL_FRect
  - SDL_FreeSurface -> SDL_DestroySurface, SDL_QueryTexture -> SDL_GetTextureSize
  - TTF_RenderText_Blended signature change (added length param)
  - Event enum renames (SDL_QUIT -> SDL_EVENT_QUIT, etc.)
  - Key access: event.key.keysym.sym -> event.key.key
- Replace ImGui SDL2 backends with SDL3 backends (v1.91.0, patched for SDL3 3.4.x)
- Remove SDL2 .lib/.dll files from repository
- Disable AVIF/JXL/WebP in SDL3_image (avoids NASM dependency)
- Add CMakePresets.json for cross-platform builds
- Update .gitignore for SDL header directories

+++ src/Components/AnimationComponent.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <SDL2/SDL.h>
+#include <SDL3/SDL.h>
 
 struct AnimationComponent
 {
@@ -8,7 +8,7 @@ struct AnimationComponent
 	uint8_t CurrentFrame;
 	uint8_t FramesPerSecond;
 	bool Loop;
-	uint32_t StartTime;
+	uint64_t StartTime;
 
 
 	AnimationComponent(uint8_t TotalFrames = 1, uint8_t FramesPerSecond = 1, bool Loop = true)
