+++ src/Components/ProjectileComponent.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <SDL2/SDL.h>
+#include <SDL3/SDL.h>
 #include <stdint.h>
 
 struct ProjectileComponent
@@ -8,7 +8,7 @@ struct ProjectileComponent
 	bool IsFriendly;
 	uint8_t HitPercentDamage;
 	uint16_t Duration;
-	uint16_t StartTime;
+	uint64_t StartTime;
 
 	ProjectileComponent(bool IsFriendly = false, uint8_t HitPercentDamage = 0, uint16_t Duration = 0)
 	{
